+++ tests/model_serving/model_runtime/triton/basic_model_deployment/test_onnx_model.py
@@ -82,7 +82,7 @@ pytestmark = pytest.mark.usefixtures(
     ],
     indirect=True,
 )
-class TestonnxModel:
+class TestONNXModel:
     """
     Test class for onnx inference using Triton on KServe.
